Improved search functionality and implementation.
